Cleaned up the the file structure once again, but this time in the C programs folder of cis231.
The next cleaing was to rename the folder for cis103 to the date that it pertains to. All projects that are created in class should be made in this fashion, and if a project is use on multiple days, then it should be copied. Any assignments will be named according to their purpose or function instead of date, but may have the date appendend to the title based on the teachers naming convention. This is all subject to change if the teacher has a specific naming convention that should be complied to.