Add integer input method.
This method grabs a line of input and then parses it with a regular
expression into integers. If a particular number doesn't fit in an
integer, then it will fail to parse, and a warning message will be
shown to the user saying that whatever the number was, won't fit
into an Integer.

The spec requires looping for input to only be interrupted in the
event that the user enters a newline, so the function ignores
non-numbers, and invalid input to just keep going.

When nothing is entered the scanner will return an empty string
because the newline is matched by the scanner's regexp.