Sort with insertion sort.
The numbers need to be sorted from high to low, and the least
effort way to do this is with insertion sort. Nothing needs to
be returned because the ArrayList is an object and therefore passed
by it's reference. So it's state is the same inside and outside of
the sorting function.