Add Lab 4.
Lab 4 is using the concepts learned in Chapter 4, such as the
abstract class and comparable. To silence the errors that this
program gives on compile time, Generics would have to have been
used when using compare to. But we had not fully covered Generics,
so I just ignored the warnings. Especially since it was explicitly
stated that we should have a Comparable[] as the parameter for our
sorting method which eliminates the possibility of using Generics.

A simple implementation of a Regular Octogon concrete class is
implemented and extends the abstract class Shape. It then gives its
appropriate methods for finding the perimeter and area. The
semiperimeter method did not need to be implemented becauese that is
done in Shape class.

We are not supposed to turn the Shape Class, so absolutely nothing
in it should be changed from the original version found in the text
book.

The main method creates an array of Regular Octogons with random
lengths and then sorts them. They are output before and after
the sorting.