Add in weiss shapes
The instructor sent out an email saying that the shape classes created
by Mark Allen Weiss were to be thrown in the package along with the rest
of our code. These shapes are the shapes that come from the code samples
that are available online. These aren't included in the book.