Create shapes with a simple if else
Creating shapes doesn't need to be that fancy because there are so few
of them. It would have been cool to do something with reflection to
dynamically check for the shape class and to dynamically check for what
needed to be given to the shape to fulfill it's constructor parameters.
But this is really complex for something with under ten shapes.

After the shape's type has been found then the constructor arguments
will be retrieved from user input. The shape will then be created and
returned. If there is no shape type matching the input, then null will
be returned.