Add Square Shape
The square shape is just similar to the other shapes that only need one
side to represent the entire shape.

There was a typo in the spec that may have suggested to make a Square
Regular Hexagon, but seeing that there is no Square Regular Hexagon, it
was just needing a comma.