Add shell system or handling user input
Because the options for this program are so limited, it makes more sense
to make a simple system of if/else statements rather than create a
complex shell program.

All features are operating but the list function doesn't properly sort
the Shapes. This is so I can focus on getting the shell to work.