Add comparators and sort
The sorts in this assignemnet are to be done by ascending descending and
either compared to as a whole by the area and perimeter, or the class
name and the area and perimeters. So far just the asceding for the class
sort is impleemented. All of the shpaes need to implement the compareTo
functions to be able to be compared and thus sorted by the sort
function.

The descending function will be similar to the asceding but simply
return the negative.