Finish up sorting.
I completely abondoned the original compareTo sorting method because
there would have to be two ways to check to sort by and it is much
quicker to implement it by just doing a compare to check after manually
retrieving the values.

Doing it any other way would create a lot of complexity and make it hard
to follow.

There are now optiosn for all of the ways that the data should be
output.

I ws debating having the shapes list their area and perimeters, but
since it wasn't explicitly stated in the spec, There was no real point
in doing it.