Add ascending sort method.
Since this collection needs to be sorted in ascending order for
most of it's lifetime, then there should be a sort method to do
such a thing. I picked insertion sort since it will be most
efficient when the the list is small and performs better if the
list is already sorted.

This method is only public now for isolated testing of the function
and it will be private when the sorting of add is done.