Add binary search
The binary search searches the sorted array and will return the
index of the target element when it is found, or if it is not found
then the negative value minus one for where it should be placed
will be returned instead. If the list has nothing in it, then it
will return -1 since 0 is where the element should be placed if
it were to be inserted and -1 less than 0.

This method is very different than the other binary search method
since it will be used for locating where an element should be
placed when add is used or set.